pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "loopflow"
version = "0.1.0"
description = "Orchestrate LLM collaboration for file generation"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "anthropic>=0.18.0",
    "openai>=1.12.0",
    "aiohttp>=3.9.0",
    "code-context>=0.1.0"
]

[project.scripts]
loopflow = "loopflow.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["loopflow"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel.force-include]
"loopflow/templates/" = "loopflow/templates"

